Move user capabilities to a mixin and include in relevant models
